Add all files from the SaaS repo
This is the script I have been working on for the past couple weeks. Obtains statistics, provides analysis, and delivers a clean SaaS proposal for what I envision the site looking like. Note that the UI / UX is not complete; I want to make the site more dynamic, alter wording, and clean some other misc. components up. This is mostly to kickstart site development and have a prototype up.